// tests/valnor-test-mssql/src/codegen/valnor_test.type_coverage-table.ts
export const TypeCoverage = valnor.newSqlTable<{ Select: ITypeCoverageSelect, Insert: ITypeCoverageInsert, Update: ITypeCoverageUpdate; Delete: true }>({
   crud:
   {
      select: true, 
      insert: true, 
      update: true, 
      delete: true, 
   },
   tableInfo:
   {
      name: "type_coverage",
      schema: "valnor_test",
   },
   pk: ["colUniqueidentifier"], 
   dialect: "tsql",
   columns:
   {

      /**
       * col_uniqueidentifier uniqueidentifier default (newid())
      */
      colUniqueidentifier: "col_uniqueidentifier",

      /**
       * col_varchar varchar
      */
      colVarchar: "col_varchar",

      /**
       * col_nvarchar nvarchar
      */
      colNvarchar: "col_nvarchar",

      /**
       * col_char char
      */
      colChar: "col_char",

      /**
       * col_nchar nchar
      */
      colNchar: "col_nchar",

      /**
       * col_text text
      */
      colText: "col_text",

      /**
       * col_ntext ntext
      */
      colNtext: "col_ntext",

      /**
       * col_xml xml
      */
      colXml: "col_xml",

      /**
       * col_time time
      */
      colTime: "col_time",

      /**
       * col_int int
      */
      colInt: "col_int",

      /**
       * col_smallint smallint
      */
      colSmallint: "col_smallint",

      /**
       * col_tinyint tinyint
      */
      colTinyint: "col_tinyint",

      /**
       * col_decimal decimal
      */
      colDecimal: "col_decimal",

      /**
       * col_numeric numeric
      */
      colNumeric: "col_numeric",

      /**
       * col_float float
      */
      colFloat: "col_float",

      /**
       * col_real real
      */
      colReal: "col_real",

      /**
       * col_money money
      */
      colMoney: "col_money",

      /**
       * col_smallmoney smallmoney
      */
      colSmallmoney: "col_smallmoney",

      /**
       * col_bigint bigint
      */
      colBigint: "col_bigint",

      /**
       * col_bit bit
      */
      colBit: "col_bit",

      /**
       * col_date date
      */
      colDate: "col_date",

      /**
       * col_datetime datetime
      */
      colDatetime: "col_datetime",

      /**
       * col_datetime2 datetime2
      */
      colDatetime2: "col_datetime2",

      /**
       * col_smalldatetime smalldatetime
      */
      colSmalldatetime: "col_smalldatetime",

      /**
       * col_datetimeoffset datetimeoffset default (sysdatetimeoffset())
      */
      colDatetimeoffset: "col_datetimeoffset",

      /**
       * col_binary binary
      */
      colBinary: "col_binary",

      /**
       * col_varbinary varbinary
      */
      colVarbinary: "col_varbinary",

      /**
       * col_image image
      */
      colImage: "col_image",
   }
});

export type ITypeCoverageInsert = {
   colUniqueidentifier?: string;
   colVarchar: string;
   colNvarchar: string;
   colChar: string;
   colNchar: string;
   colText: string;
   colNtext: string;
   colXml?: string | null;
   colTime?: Date | null;
   colInt: number;
   colSmallint: number;
   colTinyint: number;
   colDecimal?: number | null;
   colNumeric?: number | null;
   colFloat: number;
   colReal: number;
   colMoney?: number | null;
   colSmallmoney?: number | null;
   colBigint: string;
   colBit: boolean;
   colDate: Date;
   colDatetime: Date;
   colDatetime2: Date;
   colSmalldatetime: Date;
   colDatetimeoffset?: Date;
   colBinary?: Uint8Array | null;
   colVarbinary?: Uint8Array | null;
   colImage?: Uint8Array | null;
}

export type ITypeCoverageUpdate = Partial<ITypeCoverageInsert>;

export type ITypeCoverageSelect = {
   colUniqueidentifier: string
   colVarchar: string
   colNvarchar: string
   colChar: string
   colNchar: string
   colText: string
   colNtext: string
   colXml: string | null
   colTime: Date | null
   colInt: number
   colSmallint: number
   colTinyint: number
   colDecimal: number | null
   colNumeric: number | null
   colFloat: number
   colReal: number
   colMoney: number | null
   colSmallmoney: number | null
   colBigint: string
   colBit: boolean
   colDate: Date
   colDatetime: Date
   colDatetime2: Date
   colSmalldatetime: Date
   colDatetimeoffset: Date
   colBinary: Uint8Array | null
   colVarbinary: Uint8Array | null
   colImage: Uint8Array | null
}